requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0.0
click>=8.0.0
requests>=2.28.0

# Data quality
baselinr>=0.1.0

# Orchestration
dagster>=1.5.0
dagster-webserver>=1.5.0
dagster-postgres>=0.21.0

# Transformation
sqlmesh>=0.0.1

# Ingestion
dlt[postgres]>=0.3.0
psycopg2-binary>=2.9.0

# Utilities
tabulate>=0.9.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
pandas>=2.0.0
joblib>=1.3.0
shap>=0.43.0
mlflow==2.14.2
feast[postgres]>=0.38.0

# Historical Data Backfill (free, open-source)
basketball-reference-scraper>=1.0.0
beautifulsoup4>=4.12.0

# Catalog
datahub>=0.12.0

# Development dependencies (install with: pip install -e ".[dev]")
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
# isort>=5.12.0

